(02APL)_AiO_Calc
This program is still not so perfect, i hope you want to help me to perfect it. This program used for my task for final test of semester 2 in my college. I proud of myself because I was challenged of lecturer assistant. They are 2 guys, and my seniors are also in my college. I was challenged for create some complicated program. The result is i won my perfect presentation of my program. This program is helped and belongs to my group.  Thank you~